move header impl into separate module (where all impls will live)
- names are going to change and usage of mod.rs instead of cur structure
